fix(harness): restore setup_product_chain/_build_mock_context_manager to MediaBuyCreateEnv
MediaBuyCreateEnv called self.setup_product_chain() and self._build_mock_context_manager() but defined neither — they lived only on a _MediaBuyCreateEnv bridge subclass in the integration suite, so any consumer of the bare class (or MediaBuyDualEnv, which has no bridge) hit AttributeError. Move both methods onto MediaBuyCreateEnv and delete the now-empty bridge; MediaBuyDualEnv inherits them via its MRO.

Also restore two further dangling references that kept media_buy_dual.py from importing at all (make_adapter_update_side_effect in _mixins, _WRAPPER_UNSUPPORTED_FIELDS in the update harness). Lower the behavioral-mock cap 39->38 to track the mock construction that moved out of the integration suite into the harness.
