+++ tests/harness/media_buy_create.py
@@ -67,6 +67,77 @@ class MediaBuyCreateEnv(IntegrationEnv):
         product, pricing_option = self.setup_product_chain(tenant)
         return tenant, principal, product, pricing_option
 
+    def setup_product_chain(
+        self,
+        tenant: Any,
+        *,
+        product_id: str = "prod_1",
+        currency: str = "USD",
+        with_pricing: bool = True,
+        format_ids: list[dict[str, str]] | None = None,
+    ) -> tuple:
+        """Seed a real PropertyTag/Product/PricingOption row set for *tenant*.
+
+        Creates the "all_inventory" PropertyTag once (idempotent across calls),
+        a non-guaranteed Product referencing it, and (when ``with_pricing``) a
+        fixed CPM PricingOption. Returns (product, pricing_option).
+        """
+        from tests.factories import PricingOptionFactory, ProductFactory
+        from tests.factories.core import PropertyTagFactory
+
+        existing_tag = getattr(self, "_seeded_all_inventory_tag", False)
+        if not existing_tag:
+            PropertyTagFactory(tenant=tenant, tag_id="all_inventory", name="All Inventory")
+            self._seeded_all_inventory_tag = True
+
+        if format_ids is None:
+            format_ids = [{"agent_url": "https://creative.adcontextprotocol.org", "id": "display_300x250"}]
+
+        product = ProductFactory(
+            tenant=tenant,
+            product_id=product_id,
+            delivery_type="non_guaranteed",
+            format_ids=format_ids,
+            property_tags=["all_inventory"],
+        )
+        pricing_option = None
+        if with_pricing:
+            pricing_option = PricingOptionFactory(
+                product=product, pricing_model="cpm", currency=currency, is_fixed=True
+            )
+        return product, pricing_option
+
+    def _build_mock_context_manager(self, tool_name: str) -> MagicMock:
+        """Wrap the real ContextManager so create_context/create_workflow_step persist.
+
+        The manual approval path satisfies ObjectWorkflowMapping foreign keys by
+        creating real Context/WorkflowStep rows. ``get_context`` returns None and
+        the update/message hooks are no-ops.
+        """
+        from src.core.context_manager import get_context_manager
+
+        real = get_context_manager()
+        mgr = MagicMock()
+
+        def _create_context(*_args: Any, **kwargs: Any):
+            return real.create_context(
+                tenant_id=kwargs.get("tenant_id", self._tenant_id),
+                principal_id=kwargs.get("principal_id", self._principal_id),
+            )
+
+        def _create_workflow_step(*_args: Any, **kwargs: Any):
+            kwargs.setdefault("step_type", "media_buy_creation")
+            kwargs.setdefault("owner", "system")
+            kwargs.setdefault("tool_name", tool_name)
+            return real.create_workflow_step(**kwargs)
+
+        mgr.create_context.side_effect = _create_context
+        mgr.get_context.return_value = None
+        mgr.create_workflow_step.side_effect = _create_workflow_step
+        mgr.update_workflow_step.return_value = None
+        mgr.add_message.return_value = None
+        return mgr
+
     def _configure_mocks(self) -> None:
         """Set up happy-path defaults for external mocks."""
         # Adapter: mock create_media_buy — returns response matching the request packages.
