+++ tests/integration/test_create_media_buy_behavioral.py
@@ -75,88 +75,19 @@ _PRINCIPAL_ID = "behavioralprincipal"
 
 
 # ---------------------------------------------------------------------------
-# Harness subclass + request helper
+# Harness factory + request helper
 # ---------------------------------------------------------------------------
 
 
-class _MediaBuyCreateEnv(MediaBuyCreateEnv):
-    """MediaBuyCreateEnv with the product-chain and context-manager builders.
-
-    setup_product_chain seeds a real PropertyTag/Product/PricingOption row set.
-    _build_mock_context_manager wraps the real ContextManager so create_context
-    and create_workflow_step persist real Context/WorkflowStep rows — the manual
-    approval path then satisfies the ObjectWorkflowMapping foreign keys.
-    """
-
-    def setup_product_chain(
-        self,
-        tenant: Any,
-        *,
-        product_id: str = "prod_1",
-        currency: str = "USD",
-        with_pricing: bool = True,
-        format_ids: list[dict[str, str]] | None = None,
-    ) -> tuple:
-        from tests.factories import PricingOptionFactory, ProductFactory
-        from tests.factories.core import PropertyTagFactory
-
-        existing_tag = getattr(self, "_seeded_all_inventory_tag", False)
-        if not existing_tag:
-            PropertyTagFactory(tenant=tenant, tag_id="all_inventory", name="All Inventory")
-            self._seeded_all_inventory_tag = True
-
-        if format_ids is None:
-            format_ids = [{"agent_url": "https://creative.adcontextprotocol.org", "id": "display_300x250"}]
-
-        product = ProductFactory(
-            tenant=tenant,
-            product_id=product_id,
-            delivery_type="non_guaranteed",
-            format_ids=format_ids,
-            property_tags=["all_inventory"],
-        )
-        pricing_option = None
-        if with_pricing:
-            pricing_option = PricingOptionFactory(
-                product=product, pricing_model="cpm", currency=currency, is_fixed=True
-            )
-        return product, pricing_option
-
-    def _build_mock_context_manager(self, tool_name: str) -> MagicMock:
-        from src.core.context_manager import get_context_manager
-
-        real = get_context_manager()
-        mgr = MagicMock()
-
-        def _create_context(*_args: Any, **kwargs: Any):
-            return real.create_context(
-                tenant_id=kwargs.get("tenant_id", self._tenant_id),
-                principal_id=kwargs.get("principal_id", self._principal_id),
-            )
-
-        def _create_workflow_step(*_args: Any, **kwargs: Any):
-            kwargs.setdefault("step_type", "media_buy_creation")
-            kwargs.setdefault("owner", "system")
-            kwargs.setdefault("tool_name", tool_name)
-            return real.create_workflow_step(**kwargs)
-
-        mgr.create_context.side_effect = _create_context
-        mgr.get_context.return_value = None
-        mgr.create_workflow_step.side_effect = _create_workflow_step
-        mgr.update_workflow_step.return_value = None
-        mgr.add_message.return_value = None
-        return mgr
-
-
-def _env(**overrides: Any) -> _MediaBuyCreateEnv:
+def _env(**overrides: Any) -> MediaBuyCreateEnv:
     """Build the harness with a hyphen-safe tenant and explicit approval flags."""
     overrides.setdefault("tenant_id", _TENANT_ID)
     overrides.setdefault("principal_id", _PRINCIPAL_ID)
     overrides.setdefault("human_review_required", False)
-    return _MediaBuyCreateEnv(**overrides)
+    return MediaBuyCreateEnv(**overrides)
 
 
-def _require_manual_approval(env: _MediaBuyCreateEnv) -> None:
+def _require_manual_approval(env: MediaBuyCreateEnv) -> None:
     """Make the mock adapter opt into manual approval for create_media_buy.
 
     The approval branch needs both tenant.human_review_required (set via the env
