+++ tests/unit/test_architecture_behavioral_mock_cap.py
@@ -32,7 +32,7 @@ from pathlib import Path
 # Frozen at the current count; can only shrink. New behavioral files with mock
 # constructions fail immediately (force a deliberate cap entry or harness use).
 BEHAVIORAL_MOCK_CONSTRUCTION_CAP: dict[str, int] = {
-    "tests/integration/test_create_media_buy_behavioral.py": 39,
+    "tests/integration/test_create_media_buy_behavioral.py": 38,
     "tests/integration/test_creative_sync_behavioral.py": 2,
     "tests/integration/test_delivery_poll_behavioral.py": 2,
     "tests/integration/test_delivery_webhook_behavioral.py": 5,
